move promise to own module
add `Read`, `BufRead` impls for `ReadStream`
