+++ hydro_deploy/core/Cargo.toml
@@ -37,6 +37,6 @@ serde_path_to_error = "0.1.0"
 shell-escape = "0.1.0"
 tempfile = "3.0.0"
 tokio = { version = "1.29.0", features = ["full"] }
-tokio-stream = { version = "0.1.3", default-features = false }
+tokio-stream = { version = "0.1.3", default-features = false, features = ["io-util"] }
 tokio-util = { version = "0.7.5", features = ["compat", "io-util"] }
 wholesym = "0.8.1"
